+++ hi_snex/snex_core/snex_jit_BaseScope.h
@@ -189,7 +189,7 @@ public:
 	void enlargeAllocatedSize(const TypeInfo& t)
 	{
 		jassert(data == nullptr);
-		allocatedSize += t.getRequiredByteSize() + t.getRequiredAlignment();
+		allocatedSize += (int)(t.getRequiredByteSize() + t.getRequiredAlignment());
 	}
 
 	void finalise()
@@ -216,7 +216,7 @@ public:
 			auto numToPad = alignment - unAligned;
 
 			memset(data + numUsed, 0, numToPad);
-			numUsed += numToPad;
+			numUsed += (int)numToPad;
 			return true;
 		}
 
@@ -295,7 +295,7 @@ public:
 		newEntry.s = s;
 		newEntry.data = data + numUsed;
 
-		numUsed += size;
+		numUsed += (int)size;
 		memset(newEntry.data, 0, size);
 
 		jassert(numUsed <= allocatedSize);
